+++ app_manager_plugin/src/app_manager_plugin/__init__.py
@@ -1,3 +1,3 @@
-from app_manager_plugin.app_manager_plugin import AppManagerPlugin  # NOQA
+from app_manager_plugin.app_manager_plugin_base import AppManagerPlugin  # NOQA
 from app_manager_plugin.test_start_plugin import TestStartPlugin  # NOQA
 from app_manager_plugin.test_stop_plugin import TestStopPlugin  # NOQA
