+++ app_manager_plugin/src/app_manager_plugin/app_manager_plugin_base.py
